add convert/
This is Python, to help manage the conversion of old MTS, CR2, and large
JPG images. The goal is to move to AVIF and AV1, open formats. There is
a lot that is not documented here, but, it's working pretty well.

Maybe these are a project of their own. BUT, it could be that these
tools are just generally helpful.

One quirk which may not be obvious: I have fully distinct SOURCE and
TARGET directories. Why bother? We never want to end up w/ a script that
eats itself. There is idempotency here, where we can pick up where we
last left off.