improve AV1 encoding.
Looking at the original file, it may make sense to keep the existing
audio, as in the case for ac3 streams. In other cases, I found PCM
audio, which is much too much.

I'm not sure I have the conditions here quite right. With more samples
and more considerations, this can expand.

In an intermediate state, if I did not account for what I was seeing,
rather than fall back, I raised an Exception. So we can do that instead,
generally.

One future thought, which may get lost, is to replace the current check
for the target file existing, and instead key off of missing EXIF, for
example.