add skip check. Minor edits.
I'm not sure about the nostdin. I saw this a few times, not sure why I'd
need it now and not before.

Fix a bad type hint. Doesn't really matter, but it's easy to fix.

Weird attempt to look for 10 bit videos. These do not re-encode well.
The string matching is not great, can't see a better way, though. Holler
if we have some better examples.

Rather than a simple exists, I'm looking to see if we have a file that
has some content in it. This is more intensive, but feels worthwhile. We
may extend this such that CR2 files have the right EXIF attributes set
-- it's possible for the first export to go through and the second to be
missed, because there are separate commands.