add MKV. Modify condition.
I found a file encoded by Instagram which had no bit_rate defined. It
was a pretty low bit-rate. So, we're saying that if it's not defined, go
to a MIN by default, rather than the MAX. This will copy over more
streams than re-encoding, which seems right. Even in the worst case,
audio is generally much smaller.