smarter skip, w/ warnings, for video.
We could do this for any format, but I'm starting off w/ video first.
This determines that the destination file is larger than the input. This
generally means we have re-encoded something which we should instead
copy over.

I think there is another / better way around this, but, this is a final
check, a separate final run, with commands that can be copied and
executed, if so desired.