basic User updates.
Adding in _middleware is pretty neat -- we can guard a whole set of
routes, and anything in its subdirectories, in one shot. It's possible
to have this as `(u)` rather than `u` to then NOT add an extra path to
the URL. This is fine, though.

I'm also still learning Fresh. What I had was `ctx` everywhere, not
understanding that I can break apart that object a bit. It MAY not even
be worth it, but I was getting tired of typing `ctx.` all the time.

I found out I'm storing everything in the Session, which is good because
that's in the DB and can be pulled out easily. It's weird because now we
have more data in the User object, like a Set of the Passkey IDs. I also
had to do something weird re: TypeScript, because it thought I didn't
have a Session when I definitely had one. I'm thinking there's a better
way of handling this.