reworked types.
re: the note I made at the end of the last commit, I went on a journey.
The types I created were oriented around NOT storing the key in the
value itself, within Deno KV. Autocomplete made this work well enough,
but it's still more typing, and not that much of a value-add. I also
realized that the key is itself not as easily extractable as we'd
expect: it contains the rest of the prefix. So, keeping the `id` in the
value, that's just the better play.

Next, I found a way to have more than one `define`. This means the type
checking we need to do on Authenticated routes are more natural. We may
not need more than these two constructs, but, it does work quite well.